+++ mirtop/gff/compare.py
@@ -6,6 +6,8 @@ import os
 import pandas as pd
 
 from mirtop.gff import header
+from mirtop.gff.body import read_attributes, read_gff_line
+from mirtop.mirna.realign import read_id
 import mirtop.libs.logger as mylog
 logger = mylog.getLogger(__name__)
 
@@ -26,13 +28,11 @@ def compare(args):
         fn_out = os.path.join(args.out, "summary.txt")
         with open(fn_out, 'w') as outh:
             for fn in result:
-                for line in result[fn]['different']:
-                    print >>outh, "%s\t%s" % (fn, line)
-                for line in result[fn]['extra']:
-                    print >>outh, "%s\t%s" % (fn, line)
-                for line in result[fn]['miss']:
-                    print >>outh, "%s\t%s" % (fn, line)
-
+                print >>outh,  "sample\tidu\tseq\ttag\tsame_mirna\t%s" % "\t".join(result[fn][0][3].keys())
+                for line in result[fn]:
+                    read = read_id(line[0])
+                    acc = "\t".join([line[3][v] for v in line[3]])
+                    print >>outh, "%s\t%s\t%s\t%s\t%s\t%s" % (fn, line[0], read, line[1], line[2], acc)
 
 def read_reference(fn):
     """Read GFF into UID:Variant key:value dict"""
@@ -41,11 +41,9 @@ def read_reference(fn):
         for line in inh:
             if line.startswith("#"):
                 continue
-            cols = line.strip().split("\t")
-            attr_v = [v.strip().split(" ")[1] for v in cols[8].strip().split(";")[:-1]]
-            attr_k = [v.strip().split(" ")[0] for v in cols[8].strip().split(";")[:-1]]
-            attr = dict(zip(attr_k, attr_v))
-            srna[attr['UID']] = [attr['Variant'], cols[8]]
+            cols = read_gff_line(line)
+            attr = cols['attrb']
+            srna[attr['UID']] = [_simplify(attr['Variant']), attr]
     return srna
 
 def _compare_to_reference(fn, reference):
@@ -53,19 +51,20 @@ def _compare_to_reference(fn, reference):
     diff = list()
     extra = list()
     miss = list()
+    results = list()
     seen = 0
     seen_reference = set()
     with open(fn) as inh:
         for line in inh:
             if line.startswith("#"):
                 continue
-            cols = line.strip().split("\t")
-            cols = line.strip().split("\t")
-            attr_v = [v.strip().split(" ")[1] for v in cols[8].strip().split(";")[:-1]]
-            attr_k = [v.strip().split(" ")[0] for v in cols[8].strip().split(";")[:-1]]
-            attr = dict(zip(attr_k, attr_v))
+            cols = read_gff_line(line)
+            attr = cols['attrb']
             if attr['UID'] in reference:
-                if attr['Variant'] == reference[attr['UID']][0]:
+                mirna = "Y" if attr['Name'] == reference[attr['UID']][1]['Name'] else attr['Name']
+                accuracy =  _accuracy(_simplify(attr['Variant']), reference[attr['UID']][0])
+                results.append([attr['UID'], "D", mirna, accuracy])
+                if _simplify(attr['Variant']) == reference[attr['UID']][0]:
                     same += 1
                 else:
                     diff.append("%s | reference: %s" % (line.strip(), reference[attr['UID']][1]))
@@ -73,15 +72,21 @@ def _compare_to_reference(fn, reference):
                 seen_reference.add(attr['UID'])
             else:
                 extra.append("%s | extra" % line.strip())
+                results.append([attr['UID'], "E", attr['Name'], _accuracy(_simplify(attr['Variant']), "")])
     for uid in reference:
         if uid not in seen_reference:
+            results.append([uid, "M", "N", _accuracy("", reference[uid][0])])
             miss.append("| miss %s" %  reference[uid][1])
     logger.info("Number of sequences found in reference: %s" % seen)
     logger.info("Number of sequences matches reference: %s" % same)
     logger.info("Number of sequences different than reference: %s" % len(diff))
     logger.info("Number of sequences extra sequences: %s" % len(extra))
     logger.info("Number of sequences missed sequences: %s" % len(miss))
-    return {'different': diff, 'extra': extra, 'miss': miss}
+    return results
+
+def _simplify(variant):
+    simple = [v.split(":")[0] for v in variant.split(",")]
+    return ",".join(simple)
 
 def _get_samples(fn):
     with open(fn) as inh:
@@ -90,3 +95,23 @@ def _get_samples(fn):
                 return line.strip().split(": ")[1].strip().split(",")
     raise ValueError("%s doesn't contain COLDATA header." % fn)
 
+def _accuracy(target, reference):
+    """Compare each isomir label in Variant field
+       and return a list with values whether:
+           FP: no in reference
+           FN: no in target
+           TP: same values
+    """
+    logger.debug("COMPARE::ACCURACY::values %s vs %s" % (target, reference))
+    accuracy = dict()
+    types =  ["iso_5p", "iso_3p", "iso_add", "iso_snp",
+              "iso_snp_seed", "iso_snp_central",
+              "iso_snp_central_supp", "iso_snp_central_offset"]
+    for t in types:
+        if t in reference:
+            accuracy[t] = "TP" if t in target else "FN"
+        else:
+            accuracy[t] = "TN" if t not in target else "FP"
+    logger.debug("COMPARE::ACCURACY::%s" % accuracy.keys())
+    logger.debug("COMPARE::ACCURACY::%s" % accuracy.values())
+    return accuracy