+++ mirtop/gff/stats.py
@@ -4,8 +4,10 @@ Produce stats from GFF3 format
 
 import os
 import pandas as pd
-
 from mirtop.gff import header
+from mirtop.gff.body import read_attributes, read_gff_line
+from mirtop import version
+
 import mirtop.libs.logger as mylog
 logger = mylog.getLogger(__name__)
 
@@ -15,17 +17,23 @@ def stats(args):
     """
     From a list of files produce stats
     """
+    v = version.__version__
+    message_info = ("# mirtop stats version {v}").format(**locals())
     out = list()
     for fn in args.files:
         if not os.path.exists(fn):
             raise IOError("%s doesn't exist" %s)
+        logger.info("Reading: %s" % fn)
         out.append(_calc_stats(fn))
     df_final = pd.concat(out)
     outfn = os.path.join(args.out, "mirtop_stats.txt")
     if args.out != "tmp_mirtop":
-        df_final.to_csv(outfn)
+        with open(outfn, 'a') as outh:
+            print >>outh, message_info
+            df_final.to_csv(outh)
         logger.info("Stats saved at %s" % outfn)
     else:
+        print message_info
         print df_final
 
 def _get_samples(fn):
@@ -41,16 +49,19 @@ def _calc_stats(fn):
     """
     samples = _get_samples(fn)
     lines = []
+    seen = set()
     with open(fn) as inh:
         for line in inh:
             if line.startswith("#"):
                 continue
             cols = line.strip().split("\t")
-            logger.debug("## STATS: attribute %s" % cols[8])
-            attr_v = [v.strip().split(" ")[1] for v in cols[8].strip().split(";")[:-1]]
-            attr_k = [v.strip().split(" ")[0] for v in cols[8].strip().split(";")[:-1]]
-            attr = dict(zip(attr_k, attr_v))
-            lines.extend(_classify(cols[2], attr, samples))
+            cols = read_gff_line(line)
+            logger.debug("## STATS: attribute %s" % cols['attrb'])
+            attr = cols['attrb']
+            if "-".join([attr['Variant'], attr['Name']]) in seen:
+                continue
+            seen.add("-".join([attr['Variant'], attr['Name']]))
+            lines.extend(_classify(cols['type'], attr, samples))
     df = _summary(lines)
     return df